Views updated

+++ NewBlogProject.Services/Concrete/ArticleService.cs
@@ -27,9 +27,13 @@ namespace NewBlogProject.Services.Concrete
             article.DeletedDate = DateTime.Now;
             _unitOfWork.Repository<Article>().Delete(article);
 
-            article.Picture.DeletedDate = DateTime.Now;
-            _unitOfWork.Repository<Picture>().Delete(article.Picture.Id);
-
+            if (article.PictureId != null)
+            {
+                Guid id = article.Picture.Id;
+                Picture articlePicture = _unitOfWork.Repository<Picture>().FindById(id);
+                articlePicture.DeletedDate = DateTime.Now;
+                _unitOfWork.Repository<Picture>().Delete(id);
+            }
             return _unitOfWork.SaveChanges() > 0;
         }
 
@@ -37,7 +41,7 @@ namespace NewBlogProject.Services.Concrete
         {
             return _unitOfWork.Repository<Article>().FindById(id);
         }
-    
+
         public Article Insert(Article article, HttpPostedFileBase uploadPicture)
         {
 
@@ -47,7 +51,7 @@ namespace NewBlogProject.Services.Concrete
                 var newFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(uploadPicture.FileName);
                 _fileUploadManager.UploadFile(uploadPicture, newFileName, "~/Images/ArticlePicture");
 
-                
+
                 article.Picture = new Picture() // Farklı olsun :)
                 {
                     Id = Guid.NewGuid(),
@@ -57,7 +61,7 @@ namespace NewBlogProject.Services.Concrete
                 };
 
 
-            _unitOfWork.Repository<Picture>().Insert(article.Picture);
+                _unitOfWork.Repository<Picture>().Insert(article.Picture);
                 _unitOfWork.SaveChanges();
                 article.PictureId = article.Picture.Id;
             }
@@ -70,44 +74,51 @@ namespace NewBlogProject.Services.Concrete
 
         public IEnumerable<Article> Select()
         {
-            var article= _unitOfWork.Repository<Article>().Select().OrderBy(a => a.CreatedDate);  
+            var article = _unitOfWork.Repository<Article>().Select().Where(a => a.DeletedDate == null).OrderByDescending(a => a.CreatedDate);
             return article;
         }
 
         public bool Update(Article article, HttpPostedFileBase uploadPicture)
         {
             Article newArticle = _unitOfWork.Repository<Article>().FindById(article.Id);
+
             if (uploadPicture != null)
             {
-                var id = Guid.Parse(Convert.ToString(newArticle.PictureId));
-                Picture picture = _unitOfWork.Repository<Picture>().FindById(id);
-
-                if (System.IO.File.Exists(HttpContext.Current.Server.MapPath("~/Images/ArticlePicture/" + picture.Id)))
+                if (newArticle.PictureId!=null)
                 {
-                    System.IO.File.Delete(HttpContext.Current.Server.MapPath("~/Images/" + picture.Path));
+                    var id = Guid.Parse(Convert.ToString(newArticle.PictureId));
+                    Picture picture = _unitOfWork.Repository<Picture>().FindById(id);
+
+                    if (System.IO.File.Exists(HttpContext.Current.Server.MapPath("~/Images/ArticlePicture/" + picture.Id)))
+                    {
+                        System.IO.File.Delete(HttpContext.Current.Server.MapPath("~/Images/" + picture.Path));
+                    }
+                    picture.DeletedDate = DateTime.Now;
+                    _unitOfWork.Repository<Picture>().Delete(picture);
+                    _unitOfWork.SaveChanges();
                 }
-                picture.DeletedDate = DateTime.Now;
-                _unitOfWork.Repository<Picture>().Delete(picture);
-                _unitOfWork.SaveChanges();
-
-                var fileName = Path.GetFileName(uploadPicture.FileName);
-                var newFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(uploadPicture.FileName);
-                _fileUploadManager.UploadFile(uploadPicture, newFileName, "~/Images/ArticlePicture");
-                Picture newPicture = new Picture()
-                {
-                    Id = Guid.NewGuid(),
-                    Path = newFileName,
-                    Title = fileName,
-                    CreatedDate = DateTime.Now
-                };
-                _unitOfWork.Repository<Picture>().Insert(newPicture);
-                _unitOfWork.SaveChanges();
-
-
-                newArticle.PictureId = newPicture.Id;
+              
+                    var fileName = Path.GetFileName(uploadPicture.FileName);
+                    var newFileName = Guid.NewGuid().ToString() + "_" + Path.GetFileName(uploadPicture.FileName);
+                    _fileUploadManager.UploadFile(uploadPicture, newFileName, "~/Images/ArticlePicture");
+                    Picture newPicture = new Picture()
+                    {
+                        Id = Guid.NewGuid(),
+                        Path = newFileName,
+                        Title = fileName,
+                        CreatedDate = DateTime.Now
+                    };
+                    _unitOfWork.Repository<Picture>().Insert(newPicture);
+                    _unitOfWork.SaveChanges();
+                    newArticle.PictureId = newPicture.Id;
             }
-            newArticle.UpdatedDate = DateTime.Now;
 
+            newArticle.UpdatedDate = DateTime.Now;
+            newArticle.Text = article.Text;
+            newArticle.Title = article.Title;
+            newArticle.CategoryId = article.CategoryId;
+            newArticle.Description = article.Description;
+            newArticle.IsActive = article.IsActive;
 
             _unitOfWork.Repository<Article>().Update(newArticle);
             return _unitOfWork.SaveChanges() > 0;
